web_app/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
Pillow>=9.5.0
matplotlib>=3.7.0
gradio>=4.0.0
plotly>=5.14.0
pandas>=2.0.0
